+++ examples/ExampleUseDeviceNested.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * @author Maksim Naumov <me@yukki.name>
+ * @link http://yukki.name/
+ *
+ * @version 1.0.0
+ *
+ * GitHub Repo: @link https://github.com/fromYukki/Yii-MongoDB-Driver
+ * Issues: @link https://github.com/fromYukki/Yii-MongoDB-Driver/issues
+ * Documentation: @link https://github.com/fromYukki/Yii-MongoDB-Driver/wiki
+ */
+
+class ExampleUseDeviceNested extends YMongoModel
+{
+    /** @var string */
+    public $id;
+    /** @var string */
+    public $name;
+
+    /**
+     * Returns the validation rules for attributes.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return array(
+            // Required fields
+            array('id', 'required'),
+        );
+    }
+}