+++ src/YMongoCriteria.php
@@ -1,8 +1,16 @@
 <?php
 
 /**
- * CDbCriteria analogue for MongoDB. Use this class is not required, but may be useful.
+ * @author Maksim Naumov <me@yukki.name>
+ * @link http://yukki.name/
+ *
+ * @version 1.0.0
+ *
+ * GitHub Repo: @link https://github.com/fromYukki/Yii-MongoDB-Driver
+ * Issues: @link https://github.com/fromYukki/Yii-MongoDB-Driver/issues
+ * Documentation: @link https://github.com/fromYukki/Yii-MongoDB-Driver/wiki
  */
+
 class YMongoCriteria extends CComponent
 {
     /**