+++ src/YMongoHttpSessions.php
@@ -1,5 +1,16 @@
 <?php
 
+/**
+ * @author Maksim Naumov <me@yukki.name>
+ * @link http://yukki.name/
+ *
+ * @version 1.0.0
+ *
+ * GitHub Repo: @link https://github.com/fromYukki/Yii-MongoDB-Driver
+ * Issues: @link https://github.com/fromYukki/Yii-MongoDB-Driver/issues
+ * Documentation: @link https://github.com/fromYukki/Yii-MongoDB-Driver/wiki
+ */
+
 class YMongoHttpSessions extends CHttpSession
 {
     /**