+++ src/YMongoModel.php
@@ -1,6 +1,17 @@
 <?php
 
 /**
+ * @author Maksim Naumov <me@yukki.name>
+ * @link http://yukki.name/
+ *
+ * @version 1.0.0
+ *
+ * GitHub Repo: @link https://github.com/fromYukki/Yii-MongoDB-Driver
+ * Issues: @link https://github.com/fromYukki/Yii-MongoDB-Driver/issues
+ * Documentation: @link https://github.com/fromYukki/Yii-MongoDB-Driver/wiki
+ */
+
+/**
  * WARNING! Do not inherit from this class documents stored in the database.
  * This class can be used to sub documents.
  */