+++ src/YMongoSort.php
@@ -1,8 +1,16 @@
 <?php
 
 /**
- * This is only ever used in conjunction with CGridView and CListView. It is not designed to be used independently
+ * @author Maksim Naumov <me@yukki.name>
+ * @link http://yukki.name/
+ *
+ * @version 1.0.0
+ *
+ * GitHub Repo: @link https://github.com/fromYukki/Yii-MongoDB-Driver
+ * Issues: @link https://github.com/fromYukki/Yii-MongoDB-Driver/issues
+ * Documentation: @link https://github.com/fromYukki/Yii-MongoDB-Driver/wiki
  */
+
 class YMongoSort extends CSort
 {
     /**