+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 Please update this document each time you close an issue by your commit.
 
 ## Yii-MongoDB-Driver latest development alpha
-
+- **(enh)** Added examples (fromYukki)
 
 ## Yii-MongoDB-Driver version 1.0.0
 - **(enh)** Initial version after tons of fixes (fromYukki)